+++ src/niprefs/utils.nim
@@ -1,33 +1,15 @@
 import std/[strutils, sequtils, tables, os]
 import prefsnode
 
-proc splitDir(path: string): seq[string] =
-  ## Splits a directory path, normalizing, expanding the user and adding mount moint.
-  var path = path.normalizedPath() # Remove redundant separators (A//B, A/B/, A/./B and A/foo/../B become A/B)
-  path = path.expandTilde() # Convert "~" into the home user directory
-  result = path.split(DirSep)
-
-  if result[0] == "": # For unix systems
-    result[0] = "/" # Add mount point if it got removed when spliting the result
-
-  elif result[0].toLowerAscii == "c:":
-    result[0] = r"C:\"
-
-proc checkPath*(path: string) =
-  ## Given a file path creates all the directories until the file if they don't exist.
-  if fileExists(path):
-    return
-
-  let dirSeq = splitDir(splitFile(path).dir)
-  var path = dirSeq[0]
-
-  for e, dir in dirSeq:
-    if dir.toLowerAscii notin [r"c:\", "/"]: # On Windows C:\ and / on Linux
-      if e > 0:
-        path = joinPath(path, dir) # joinPath("trial", "sub") -> "trial/sub" (on Windows "trial\sub")
-
-      if not dirExists(path):
-        createDir(path)
+proc checkFile*(path: string) = 
+  ## Iterate through `path.parentDir.parentDirs` from the root creating all the directories that do not exist.
+  ## **Example:**
+  ## ```nim
+  ## checkFile("a/b/c") # Takes c as a file, not as a directory
+  ## checkFile("a/b/c/d.png") # Only creates a/b/c directories
+  ## ```
+  for dir in path.parentDir.parentDirs(fromRoot=true):
+    discard existsOrCreateDir(dir)
 
 proc changeNested*(table: PObjectType, keys: varargs[string], val: PrefsNode): PObjectType =
   ## Changes nested `keys` for `val` in `table`.